+++ test/pages/driver/conn_test.exs
@@ -29,7 +29,13 @@ defmodule Pages.Driver.ConnTest do
 
       assert_receive {:page_controller, :submit, :ok, params}
       assert Map.keys(params) == ~w[_csrf_token form]
-      assert params["form"] == %{"string_value" => "initial", "select_value" => "initial", "bool_value" => "false"}
+
+      assert params["form"] == %{
+               "string_value" => "initial",
+               "select_value" => "initial",
+               "bool_value" => "false",
+               "radio_value" => "initial"
+             }
     end
 
     test "handles non-redirect error renders", %{conn: conn} do
@@ -41,7 +47,13 @@ defmodule Pages.Driver.ConnTest do
 
       assert_receive {:page_controller, :submit, :error, params}
       assert Map.keys(params) == ~w[_csrf_token form]
-      assert params["form"] == %{"string_value" => "", "select_value" => "", "bool_value" => "false"}
+
+      assert params["form"] == %{
+               "string_value" => "",
+               "select_value" => "",
+               "bool_value" => "false",
+               "radio_value" => nil
+             }
     end
   end
 
@@ -62,7 +74,13 @@ defmodule Pages.Driver.ConnTest do
 
       assert_receive {:page_controller, :submit, :ok, params}
       assert Map.keys(params) == ~w[_csrf_token form]
-      assert params["form"] == %{"string_value" => "updated", "select_value" => "", "bool_value" => "false"}
+
+      assert params["form"] == %{
+               "string_value" => "updated",
+               "select_value" => "",
+               "bool_value" => "false",
+               "radio_value" => nil
+             }
     end
 
     test "handles non-redirect error renders", %{conn: conn} do
@@ -73,7 +91,13 @@ defmodule Pages.Driver.ConnTest do
 
       assert_receive {:page_controller, :submit, :error, params}
       assert Map.keys(params) == ~w[_csrf_token form]
-      assert params["form"] == %{"string_value" => "", "select_value" => "", "bool_value" => "false"}
+
+      assert params["form"] == %{
+               "string_value" => "",
+               "select_value" => "",
+               "bool_value" => "false",
+               "radio_value" => nil
+             }
     end
   end
 
@@ -118,7 +142,12 @@ defmodule Pages.Driver.ConnTest do
                |> Hq.find("#form")
                |> Hq.form_fields()
 
-      assert form_params == %{select_value: "initial", string_value: "initial", bool_value: "false"}
+      assert form_params == %{
+               select_value: "initial",
+               string_value: "initial",
+               bool_value: "false",
+               radio_value: "initial"
+             }
 
       page = page |> Pages.update_form("#form", :form, select_value: "updated")
 
@@ -127,7 +156,7 @@ defmodule Pages.Driver.ConnTest do
                |> Hq.find("#form")
                |> Hq.form_fields()
 
-      assert form_params == %{select_value: "updated", bool_value: "false"}
+      assert form_params == %{select_value: "updated", bool_value: "false", radio_value: nil}
     end
 
     test "updates checkbox inputs on a form", %{conn: conn} do
@@ -141,7 +170,12 @@ defmodule Pages.Driver.ConnTest do
                |> Hq.find("#form")
                |> Hq.form_fields()
 
-      assert form_params == %{select_value: "initial", string_value: "initial", bool_value: "false"}
+      assert form_params == %{
+               select_value: "initial",
+               string_value: "initial",
+               bool_value: "false",
+               radio_value: "initial"
+             }
 
       page = page |> Pages.update_form("#form", :form, bool_value: true)
 
@@ -150,7 +184,35 @@ defmodule Pages.Driver.ConnTest do
                |> Hq.find("#form")
                |> Hq.form_fields()
 
-      assert form_params == %{select_value: "", bool_value: "true"}
+      assert form_params == %{select_value: "", bool_value: "true", radio_value: nil}
+    end
+
+    test "updates radio inputs on a form", %{conn: conn} do
+      page =
+        conn
+        |> Pages.visit("/pages/form")
+        |> assert_success()
+
+      assert %{_csrf_token: _, form: form_params} =
+               page
+               |> Hq.find("#form")
+               |> Hq.form_fields()
+
+      assert form_params == %{
+               select_value: "initial",
+               string_value: "initial",
+               bool_value: "false",
+               radio_value: "initial"
+             }
+
+      page = page |> Pages.update_form("#form", :form, radio_value: "updated")
+
+      assert %{_csrf_token: _, form: form_params} =
+               page
+               |> Hq.find("#form")
+               |> Hq.form_fields()
+
+      assert form_params == %{radio_value: "updated", bool_value: "false", select_value: ""}
     end
   end
 end